+++ src/common/appConstants.ts
