+++ src/utils/printConsoleLog.ts
