+++ src/v1/modules/seller/models/sellers.ts
@@ -51,15 +51,14 @@ export const SellerSchema = new mongoose.Schema({
     },
   },
   store: {
-    name: { type: String, required: true },
-    location: { type: String, required: true },
+    name: { type: String, required: false },
+    location: { type: String, required: false },
     gstNumber: {
       type: String,
-      required: [true, "GST number is required"],
-      unique: true,
+      required: false,
       match: [/^[0-9]{2}[A-Z]{5}[0-9]{4}[A-Z]{1}[1-9A-Z]{1}Z[0-9A-Z]{1}$/, "Please fill a valid GST number"],
     },
-    contactDetails: { type: String, required: true },
-    timing: { type: String, required: true },
+    contactDetails: { type: String, required: false },
+    timing: { type: String, required: false },
   },
 }, { timestamps: true });
